// WebApi/ClientApp/src/models/responses/hoa-don/IHoaDonDangKyPhatHanh.ts
export interface IHoaDonDangKyPhatHanh {
    id: number;
    donvi_ma_dv: string;
    mau_so: string;
    so_luong: number;
    so_bat_dau: string;
    so_ket_thuc: string;
    ngay_su_dung: string;
    so_qd: string;
    ngay_qd: string;
    loai_hoa_don_ct_id: number;
    ky_hieu: string;
    ten_hoa_don: string;
    hoa_don_dang_ky_phat_hanh_trang_thai_id: number;
    hinh_thuc_code: string;
    is_chiu_thue: boolean
}